+++ src/pages/Dashboard/index.tsx
@@ -1,12 +1,22 @@
-import Header from "@nepMeds/components/Header/Header";
+import { Box, Flex } from "@chakra-ui/react";
+
+import Sidebar from "@nepMeds/components/Sidebar/Sidebar";
+import Navbar from "@nepMeds/components/Navbar/Navbar";
+import DashboardBody from "@nepMeds/components/Dashboard/DashboardBody";
 
 const Dashboard = () => {
   return (
     <>
-      <Header />
-      Welcome to Dashboard!!
+      <Flex>
+        <Box flex={"1"}>
+          <Sidebar />
+        </Box>
+        <Box flex={"5"}>
+          <Navbar />
+          <DashboardBody />
+        </Box>
+      </Flex>
     </>
   );
 };
-
 export default Dashboard;