+++ src/components/Table/Doctor/RegisteredDocList.tsx
@@ -1,38 +1,38 @@
 import { Badge, Icon } from "@chakra-ui/react";
+import { DataTable } from "@nepMeds/components/DataTable";
 import { useDoctorList } from "@nepMeds/service/nepmeds-doctorlist";
+import { CellContext } from "@tanstack/react-table";
 import React from "react";
 import { Show } from "react-iconly";
-import { CellProps } from "react-table";
-import ComponentTable from "../ComponentTable";
 
 const RegisteredDocList = () => {
   const columns = React.useMemo(
     () => [
       {
-        Header: "S.N",
-        accessor: "id",
+        header: "S.N",
+        accessorKey: "id",
       },
       {
-        Header: "Doctor's Name",
-        accessor: "full_name",
+        header: "Doctor's Name",
+        accessorKey: "full_name",
       },
       {
-        Header: "Contact Number",
-        accessor: "contact",
+        header: "Contact Number",
+        accessorKey: "contact",
       },
       {
-        Header: "Specialization",
-        accessor: "specialization",
-        Cell: ({ row }: CellProps<{ specialization: any }>) => {
+        header: "Specialization",
+        accessorKey: "specialization",
+        Cell: ({ row }: CellContext<{ specialization: any }, any>) => {
           const { name } = row?.original?.specialization[0] ?? "";
 
           return <p>{name}</p>;
         },
       },
       {
-        Header: "Status",
-        accessor: "status",
-        Cell: ({ row }: CellProps<{ status: string }>) => {
+        header: "Status",
+        accessorKey: "status",
+        Cell: ({ row }: CellContext<{ status: string }, any>) => {
           const { status } = row.original;
 
           return (
@@ -43,8 +43,8 @@ const RegisteredDocList = () => {
         },
       },
       {
-        Header: "Actions",
-        accessor: "actions",
+        header: "Actions",
+        accessorKey: "actions",
         Cell: () => {
           return (
             <Icon
@@ -63,7 +63,7 @@ const RegisteredDocList = () => {
   const { data, isLoading } = useDoctorList();
   return (
     <>
-      <ComponentTable columns={columns} data={data ?? []} />;
+      <DataTable columns={columns} data={data ?? []} isLoading={isLoading} />;
     </>
   );
 };